requirements.txt
openai==1.78.1
python-dotenv==1.0.1
tiktoken==0.9.0
tenacity==9.1.2
colorama==0.4.6
